moved ErrorWindow to own module and file
